[Feat #72] Set Time Graph 분리
