+++ GongdeunTop/GongdeunTop/Views/Timer/SetTime/BatchTimeSettingForm.swift
@@ -13,11 +13,9 @@ struct BatchTimeSettingForm: View {
     
     var body: some View {
         VStack {
-            Text("totalTime\(manager.getMinute(of: manager.getTotalSeconds()))")
-                .font(.title)
-                .bold()
+            SetTimeGraph(manager: manager)
             
-            drawGraph()
+            Spacer()
             
             sessionStepper
             concentrationTimeStepper
@@ -29,78 +27,7 @@ struct BatchTimeSettingForm: View {
         }
     }
 }
-//MARK: - Graph
-extension BatchTimeSettingForm {
-    private func getConcentrationTimeRatio(ofSession index: Int) -> CGFloat {
-        guard let session = manager.timeSetting.sessions[safe: index] else {
-            print("out of index")
-            return CGFloat(1)
-        }
-        let concetrationSeconds = CGFloat(session.concentrationSeconds)
-        let totalSeconds = CGFloat(manager.getTotalSeconds())
-        return concetrationSeconds / totalSeconds
-    }
-    
-    private func getRestTimeRatio(ofSession index: Int) -> CGFloat {
-        guard let session = manager.timeSetting.sessions[safe: index] else {
-            print("out of index")
-            return CGFloat(1)
-        }
-        let restSeconds = CGFloat(session.restSeconds)
-        let totalSeconds = CGFloat(manager.getTotalSeconds())
-        return restSeconds / totalSeconds
-    }
-    
-    @ViewBuilder
-    func drawGraph() -> some View {
-        GeometryReader { geo in
-            let width = geo.size.width
-            HStack(spacing: 0) {
-                ForEach(manager.timeSetting.sessions.indices, id: \.self) { index in
-                    var sessionRatio: CGFloat {
-                        getConcentrationTimeRatio(ofSession: index) + getRestTimeRatio(ofSession: index)
-                    }
-                    var sessionWidth: CGFloat {
-                        width * sessionRatio
-                    }
-                    
-                    var concentrationTimeWidth: CGFloat {
-                        width * getConcentrationTimeRatio(ofSession: index) - 4
-                    }
-                    
-                    var restTimeWidth: CGFloat {
-                        width * getRestTimeRatio(ofSession: index) - 4
-                    }
-                    
-                    HStack(spacing: 2) {
-                        Rectangle()
-                            .foregroundColor(themeManager.getColorInPriority(of: .solid))
-                            .frame(width: concentrationTimeWidth,
-                                   height: 48)
-                        
-                        Rectangle()
-                            .foregroundColor(themeManager.getColorInPriority(of: .medium))
-                            .frame(width: restTimeWidth < 0 ? 0 : restTimeWidth,
-                                   height: 48)
-                    }
-                    .padding(.horizontal, 5)
-                    .frame(width: sessionWidth)
-                    .overlay(alignment: .bottom) {
-                        Capsule()
-                            .frame(width: sessionWidth - 10, height: 5)
-                            .foregroundColor(.gray)
-                            .opacity(0.5)
-                            .offset(y: 10)
-                    }
-                }
-            }
-            .frame(width: width)
-            
-        }
-        .frame( height: 58)
-        .padding(.bottom, 40)
-    }
-}
+
 
 //MARK: - Stepper
 extension BatchTimeSettingForm {
